requirements.txt
# Core
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
pydantic>=2.9.0

# UI consola
rich>=13.9.0

# Utils
python-dotenv>=1.0.0

# Dev (opcional, para tests)
pytest>=8.0.0